+++ src/game/save_file.h
@@ -26,7 +26,7 @@ struct SaveBlockSignature {
 
 struct SaveFile {
     u16 coins; //amount of coins
-    u16 abilities;
+    u32 abilities;
     u8 ability_dpad[4];
     u8 level_f_flags;
     u32 flags;
@@ -192,7 +192,7 @@ void save_file_set_widescreen_mode(u8 mode);
 #endif
 void save_file_move_cap_to_default_location(void);
 
-u16 save_file_check_ability_unlocked(u8 ability_id);
+s32 save_file_check_ability_unlocked(u8 ability_id);
 void save_file_unlock_ability(u8 ability_id);
 
 void save_file_screenshot(void);
